Refactoring
Quick utilities refactoring in utils/ folder
